requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.24.3
openpyxl==3.1.2
python-dotenv==1.0.1
plotly==5.19.0
openai==1.3.0
requests>=2.31.0
aiohttp>=3.8.0